+++ app/src/main/java/network/minter/bipwallet/internal/di/InjectorsModule.java
@@ -52,6 +52,7 @@ import network.minter.bipwallet.internal.di.annotations.ServiceScope;
 import network.minter.bipwallet.security.ui.PinEnterActivity;
 import network.minter.bipwallet.services.livebalance.LiveBalanceService;
 import network.minter.bipwallet.settings.ui.PasswordChangeMigrationActivity;
+import network.minter.bipwallet.tx.ui.ExternalTransactionActivity;
 import network.minter.bipwallet.tx.ui.TransactionListActivity;
 
 /**
@@ -133,4 +134,8 @@ public interface InjectorsModule {
     @ContributesAndroidInjector
     @ActivityScope
     PinEnterActivity pinCodeActivityInjector();
+
+    @ContributesAndroidInjector
+    @ActivityScope
+    ExternalTransactionActivity externalTransactionActivityInjector();
 }