+++ app/src/main/java/network/minter/bipwallet/sending/contract/SendView.java
@@ -15,6 +15,7 @@ import network.minter.bipwallet.internal.mvp.ErrorViewWithRetry;
 import network.minter.bipwallet.sending.account.AccountSelectedAdapter;
 import network.minter.bipwallet.sending.adapters.RecipientListAdapter;
 import network.minter.bipwallet.sending.models.RecipientItem;
+import network.minter.blockchain.models.operational.ExternalTransaction;
 
 /**
  * minter-android-wallet. 2019
@@ -50,4 +51,6 @@ public interface SendView extends MvpView, ErrorViewWithRetry {
     void setPayloadChangeListener(TextWatcher listener);
     void setPayload(String payload);
     void setActionTitle(int buttonTitle);
+    @StateStrategyType(OneExecutionStateStrategy.class)
+    void startExternalTransaction(ExternalTransaction tx);
 }